plot.spVBEnKs {spCalibration}	R Documentation

plot of spVBEnKs class

Description

plot of spVBEnKs class

Usage

plot.spVBEnKs(
  x = NULL,
  para = NULL,
  IS = NULL,
  n = 100,
  title.size = 16,
  text.size = 12,
  line.size = 5
)

Arguments

x
the class from return by spVBEnKs().

para
option to provide a list of parameter estimation (eg. PIU) by spVBEnKs(). Default is NULL.

IS
option to provide posterior sample and its weight by importance sampling by spVBEnKs(). Default is NULL.

n
option to provide sample size used to estimate distribution of no close form based on Kernel method. Default is 100.

title.size
the axis title font size.

text.size
the axis text font size.

line.size
the line font size.

See Also

spVBEnKs.

[Package spCalibration version 0.1.0 Index]