predict.spVBEnKs {spCalibration}	R Documentation

Prediction of spVBEnKs class

Description

Prediction of spVBEnKs class

Usage

## S3 method for class 'spVBEnKs'
predict(
  x,
  test,
  site = NULL,
  method = c("ensemble", "mean"),
  transf = c("normal", "square", "log"),
  ncol.layout = 3
)

Arguments

x
the class from return by spVBEnKs().

test
option to provide a list form data, which contains 4 indexs, as follows: Y_ts, X_ts, Hs and City_Name, where Y_ts, X_ts and Hs are set as function spVBEnKs() and City_Name is the city name need to test.

site
option to provide a data frame which contains information of monitoring station. Default is NULL.

method
choose predictive method: ensemble or mean. Default is ensemble.

transf
choose method of data transform: normal, square, log. Default is normal

ncol.layout
option to provide the number of columns for figure layout.

See Also

spVBEnKs.

[Package spCalibration version 0.1.0 Index]