spVBEnKs {spCalibration}	R Documentation

Spatio-temporal data calibration by vaiational Bayesian and ensemblem kalman smoother

Description

Spatio-temporal data calibration by vaiational Bayesian and ensemblem kalman smoother.

Usage

spVBEnKs(
  data = NULL,
  prior = NULL,
  IS = NULL,
  para = NULL,
  true.para = NULL,
  parallel = FALSE,
  verbose = TRUE,
  verbose.VB = FALSE,
  Ensemble.size = 50,
  cs = 0.3,
  ct = 1,
  Remove.CPU.Count = 1,
  N.Chunk = 1,
  itMax = 100,
  tol.vb = 0.001,
  tol.real = 0.001
)

Arguments

data
option to provide a list form data, which contains 8 indexs, as follows: n, Nt, N.BAUs, Y_ts, X_ts, BAUs.Dist, Adj.Mat and Hs, where n is the observed spatial sample size; Nt is the temporal index length; N.BAUs is the number of grids; Y_ts is an Nt x n dimentional observed matrix. X_ts ia an p x n x Nt dimentional array. BAUs.Dist is an N.BAUs x N.BAUs dimentional distance matrix based on the grids which created by INLA. Adj.Mat is an adjacent matrix of grids with elements W_{ij} =1 if i and j is first-order neighbors, and 0 otherwise.

prior
option to provide a list. This list contains model parameters prior information: β, τ^2, θ_1. See subsequent examples for more details.

IS
option to provide a list which contains sample and weight of three parameters θ_2, K, k_0.

para
option to provide a list which provided initialize value of parameters. See subsequent examples for more details.

true.para
option to provide true parameters when need to simulation. Default is NULL.

parallel
option to whether need to parallel compute in importance sampling process. Default is FALSE.

verbose
if TRUE, progress of the parameter estimation is printed to the screen. Otherwise, nothing is printed to the screen.

verbose.VB
if TRUE, progress of the parameter estimation in VB process is printed to the screen. Otherwise, nothing about VB process is printed to the screen.

Ensemble.size
option to provide ensemble size. Default is 50.

cs
option to provide the ratio of maxmum distance used to taper in space.

ct
option to provide the temporal tapering range.

Remove.CPU.Count
option to remove the number of CPU when parallel is TRUE, to make sure the system does something else. Default is 1.

N.Chunk
option to divide into several parts for parameter estimation of importance sampling when size is large.

itMax
maximum number of iterations. Defaults to 1e2.

tol.vb
relative convergence tolerance in importance sampling process. Defaults to 1e-3.

tol.rel
relative convergence tolerance in complete parameter estimation. Defaults to 1e-3.

Details

For details on the implementation of 'spVBEnKs', see the website https://mgallow.github.io/spVBEnKs/articles/Details.html.

Value

returns class object spVBEnKs which includes:

Call
function call.

FinalEst
estimated model parameters.

ParaEst
the proce list of etimation of model parameters.

PIU
a list of parameter estimation.

IS
posterior sample and its weight by importance sampling.

Ks
posterior ensemble by TaEnKs algorithm.

loglik
final estimated loglikelihood.

Iterations
number of iterations.

maxit
maximum number of iterations.

Fitting_RMSE
the root mean squared error by fit model.

Fitting_MB
the mean bias by fit model.

Fitting_NMB
the normalized mean bias by fit model.

Fitting_NME
the normalized mean error by fit model.

Author(s)

YeWen Chen chenyw20150908@163.com

References

Berrocal, V. J.; Gelfand, A. E.; Holland, D. M. A bivariate space-time downscaler under space and time misalignment. The annals of applied statistics 2010, 4, 1942.

Berrocal, V. J.; Gelfand, A. E.; Holland, D. M. A spatio-temporal downscaler for output from numerical models. Journal of agricultural, biological, and environmental statistics 2010, 15, 176–197.

Katzfuss, M.; Stroud, J. R.; Wikle, C. K. Ensemble Kalman methods for high-dimensional hierarchical dynamic space-time models. Journal of the American Statistical Association 2019, 1–43.

Cressie, N.; Zammit-Mangion, A.; Wikle, C. K. Spatio-Temporal Statistics with R; Chapman and Hall, 2019.

Examples

set.seed(35426)
#####################################################################
#                       Create grids  by INLA
#####################################################################
grid <- ProduceGrid(Site,
                    max.edge = c(.3, .7), #0.3,0.7
                    offset = c(.4, .6), #0.4, 0.6
                    cutoff =.1, #0.1
                    col = "black", size = 1)
grid$plot.grid
#####################################################################
#                          Mapping matrix: H
#####################################################################
Data_Str <- ProduceHmatrix(grid, site = Site, threshold = 1e5)
#####################################################################
#                             Prior
#####################################################################
prior <- list(
               beta = list(mu = c(0, 1), Sigma2 = 1e5*diag(2))
             , nugget.tau2 = list(a = 2, b = 1)
             , theta1 = list(mu = 0.005, Sigma2 =  1e5)
             )
#####################################################################
#                         Importance sampling
#####################################################################
IS.size <- 50
IS <- list(theta2_random_sample = runif(IS.size, 1e-3, 5)
         , K_random_sample = runif(IS.size, 1e-3, 15)
         , K0_random_sample = runif(IS.size, 1e-3, 15)
         , G_theta2_weight = rep(1/IS.size, IS.size)
         , G_k_weight = rep(1/IS.size, IS.size)
         , G_k0_weight = rep(1/IS.size, IS.size)
         , IS.size = IS.size
         , Thresh = c(1, 1, 1)
         )
#####################################################################
 #                      initialize  parameters
#####################################################################
initial.para <- list(
                     beta = list(E_beta = c(2, 0.5), Sigma2 = diag(2))
                   , theta1 = list(E_theta1 = 0.005, Sigma2 = 2)
                   , k = list(E_k = 5, a = 2, b = 1)
                   , k0 = list(E_k0 = 5, a = 2, b = 1)
                   , theta2 = list(mu = 1)
                   , tau2 = list(E_tau2 = 1, a = 2, b = 1)
                    )
#####################################################################
#                             Data truncation
#####################################################################
YearMonth <- c(201511, 201512, 201601)  # 201506,201507,201508 # 201511, 201512, 201601
Yts_Xts <- ParYtsXts(Model_Base_Tab, YearMonth = YearMonth)
#####################################################################
#                           Leave one city out
#####################################################################
city_num = 1
GSD = Test_Train_Fun(Site, Data_Str, Yts_Xts, city_num)
Train = GSD$Train
Test =  GSD$Test
Train$Y_ts = sqrt(Train$Y_ts)
Train$X_ts = sqrt(Train$X_ts)
Test$X_ts = sqrt(Test$X_ts)
#####################################################################
#                           execution model
#####################################################################
CV <- spVBEnKs(data = Train, prior = prior, IS = IS,
               para = initial.para, true.para = NULL,
               parallel = TRUE, verbose = TRUE,
               verbose.VB = TRUE, Ensemble.size = 50,
               cs = 0.4, ct = 1,
               Remove.CPU.Count = 2, N.Chunk = 1,
               itMax = 1e2, tol.vb = 1e-2,
               tol.real = 1e-3)
######################################################################
#                           output
######################################################################
print(CV)
######################################################################
#                         prediction
######################################################################
# load("./data/Generate_Data/Test/CV.Rdata")
P <- predict(x = CV, test = Test, site = Site,
             method = c("ensemble"),  # ensemble, mean
             transf = "square", ncol.layout = 2)
P$error
P$ensemble.plot
######################################################################
#                      the distributions of parameters
######################################################################
plot <- plot(x = CV, para = NULL, IS = NULL, n = 500, title.size = 12,
              text.size = 8, line.size = 1)
plot$Beta.plot
plot$Close.Dist
plot$NoClose.Dist

[Package spCalibration version 0.1.0 Index]